[Thread] enable handling of ICMPv6 Echo Request
This commit reverts aosp/2954927 which disabled Echo Request handling in
OT and delegated it to the platform / host. For Android, this doesn't work well because Android will always select the same interface the
Echo Request was received to send the reply and it requires the route for the dest address to exist on the interface. Unfortunately, this
doesn't work for the case that a ICMPv6 ping is sent from a Wi-Fi device on the AIL to the OMR address of this Android device, because there is no route to the Wi-Fi device address over the Thread tunnel
interface.

This commit workaround this issue by enabling the OT Echo Request
handler. Note the original duplicate Echo Reply issue in b/324156875 has already been eliminated by aosp/3088225 because the Echo Request
will be filtered out when it's being passed to the host.

Note this issue is not limited to only ICMPv6 messages but also
general UDP or TCP messages. But given there are currently no use
cases of advertising the OMR address of the android device on AIL,
thereshouldn't be applications trying to reach to the Thread OMR
address.

Test: manually verified that ping works from a Wi-Fi device to an
      Android phone
Bug: 432438085
Change-Id: I07b32fd23c74801cc1e9760e6bb4a61ca0f8de2a