Change Backbone Router enablement.
The Backbone Router is always enabled when configuring Border Router as enabled, regardless of whether the infra link interface is ready. This causes certification test failure like 5_10_14 because the backbone agent doesn't have the correct interface. The change makes Backbone Router only start when the infra link interface is ready.

Bug: 443655457
Test: Start KW and from the log BBr started successfully without the infra if issue: https://paste.googleplex.com/5864941101907968
Change-Id: I404d37960e86a95c54335e2f0d65a99b45372437